Use non-deprecated FileHandle methods if available
We've been seeing crashes in the parts of the code which use the
deprecated `NSFileHandle` methods. These methods can raise exceptions,
but they are not `throwing`, which might explain the crashes.

Important note: this does not really solve the source of the errors, it
merely makes a better attempt at logging them properly and not crashing.
The reason these errors are thrown is not known, perhaps threading
issues?